requirements.txt
langgraph
colorama
termcolor
requests
python-dotenv
tenacity
pyautogui
playwright
uiautomation
pygetwindow
pillow
nest_asyncio
ratelimit
mainContentExtractor
pyaudio
keyboard
pyqt6
google-cloud-speech